refactor: Add Morse Code Translator program
The commit adds a Python program that translates text into Morse code and vice versa. The program uses a dictionary to represent the Morse code chart and provides functions to encrypt text to Morse code and decrypt Morse code to text. The program also includes a main function that allows the user to choose between text to Morse code conversion and Morse code to text conversion.